% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{birthday}
\usage{
birthday(n)
}
\arguments{
\item{n}{An integer or vector of integers giving the group size(s).}
}
\value{
A numeric vector of probabilities, each between 0 and 1,
corresponding to the input group sizes.
}
\description{
birthday
}
\examples{
#' # Probability for a single class size
birthday(23)   # famous case, ~0.507

# Probabilities for a range of class sizes
birthday(20:25)

# For n > 365, probability is always 1
birthday(400)
}
