% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getmode}
\alias{getmode}
\title{getmode function}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector}
}
\value{
returns the mode of the vector
}
\description{
getmode function
}
\examples{
getmode(c(1,2,2,2,3,4,5,5,6))
}
