% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterates}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
number of success
}
\description{
mybin
}
\examples{
mybin(n = 10, p = 0.7)
}
