% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample vector}

\item{fun}{funtion that is used}

\item{alpha}{percentage interval}

\item{cx}{char expansion factor}

\item{...}{}
}
\value{
a list of ci, fun, and x
}
\description{
myboot2
}
\examples{
myboot2(iter = 10000, x = c(0,1,1,1,0,0,0,1), alpha = 0.05)
}
