% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of test}

\item{iter}{number of iterations}

\item{lambda}{lambda = mean}

\item{...}{extra}
}
\value{
returns 3 graphs: hist, box, and curve
}
\description{
mycltp
}
\examples{
mycltp(n = 30, iter = 10000, lambda = 10)
}
