% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation via Grid Search}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function that accepts two arguments \code{x} and \code{param}
and returns the log-likelihood contribution for a single observation. This
function must be vectorized via \code{outer()} to handle all combinations of
data and parameter values.}

\item{x}{A numeric vector containing the observed data.}

\item{param}{A numeric vector of parameter values over which the likelihood
should be evaluated. This forms the grid for the maximization process.}

\item{...}{Additional graphical parameters passed to \code{plot()},
such as \code{main}, \code{xlab}, or \code{ylab}.}
}
\value{
A list with the following components:
\item{i}{The index in \code{param} corresponding to the maximum likelihood estimate.}
\item{parami}{The parameter value at which the likelihood is maximized (the MLE).}
\item{yi}{The maximum value of the log-likelihood.}
\item{slope}{A numeric vector describing approximate slopes around the MLE,
  useful for verifying sign changes that indicate a maximum.}
}
\description{
\code{mymaxlik()} evaluates a log-likelihood function over a grid of parameter
values and returns the parameter that maximizes the likelihood. It produces a
plot of the log-likelihood curve and marks the maximum likelihood estimate
(MLE) visually. This function is useful when analytical maximization is not
possible and numerical methods such as grid approximation are required.
}
\examples{
# Example: MLE for p in a Binomial(10, p) model
logbin <- function(x, p) log(dbinom(x, size = 10, prob = p))
x <- c(9, 9, 1, 9, 9, 9)
param <- seq(0, 1, length = 1000)
mymaxlik(lfun = logbin, x = x, param = param,
         xlab = expression(p), main = "Binomial MLE")

}
