% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve creates a curve of an norm distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{is the mean, a real number}

\item{sigma}{is the standard deviation, a real number}

\item{a}{for P(Y < a), a real number}
}
\value{
a list of mu and sigma
}
\description{
myncurve creates a curve of an norm distribution
}
\examples{
myncurve(1,1,1)
}
