% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim.R
\name{myTsim}
\alias{myTsim}
\title{myTsim}
\usage{
myTsim(
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.5,
  legend_pos = "topright",
  ...
)
}
\arguments{
\item{n1}{sample size for each simulated sample}

\item{sigma1}{true population standard deviation sigm used to generate the random samples}

\item{mean1}{ture population mean used to generate the random samples}

\item{iter}{number of monte carlo simulations to perfom}

\item{ymax}{upper limit of the y axis for the histogram to control scaling}

\item{legend_pos}{location for the legend on the plot}

\item{...}{}
}
\value{
A list with components:
\itemize{
  \item `w` – Numeric vector of simulated t statistics (length = `iter`).
  \item `summary` – Summary statistics for `w`.
  \item `sd` – Standard deviation of `w`.
  \item `fun` – Character label `"T"`.
}
}
\description{
myTsim
}
\examples{
myTsim(iter=10000,ymax=0.45)
}
