% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{# of seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of passenger showing up}
}
\value{
list(nd,nc, N,p,gamma) nd is the optimal number of seats sold (discrete).
nc is the optimal number of seats sold (continuous)
}
\description{
ntickets
}
\examples{
ntickets(200,0.02,0.95)
}
