% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoGraph.R
\name{pareto}
\alias{pareto}
\title{paretoGraph}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{is a vector}

\item{...}{}
}
\value{
creates a pareto Graph
}
\description{
paretoGraph
}
\examples{
pareto(rep(c("None","Both","LegsO","WheelsO"),c(15,8,63,20)))
}
